RDKEMW-1061: RDK-E Add COMRPC
Reason for change: Create the interface file for HdmiCecSource Test Procedure:
Risks: low
Priority: P1

+++ apis/HdmiCecSource/IHdmiCecSource.h
@@ -79,7 +79,7 @@ namespace WPEFramework
                 // @text SendKeyPressMsgEvent
                 // @param logicalAddress: Logical address of the device
                 // @param keyCode: Key code of the key press event
-                virtual void SendKeyPressMsgEvent(const int logicalAddress /* @in */, const int keyCode /* @in */) {};
+                virtual void SendKeyPressMsgEvent(const int8_t logicalAddress /* @in */, const int32_t keyCode /* @in */) {};
             };
 
             // @json:omit
